Transaction request/consumption flow

+++ app/src/main/java/co/omisego/omgshop/pages/checkout/RedeemDialogContract.kt
@@ -8,16 +8,17 @@ package co.omisego.omgshop.pages.checkout
  */
 
 import co.omisego.omgshop.base.BaseContract
+import java.math.BigDecimal
 
 interface RedeemDialogContract {
     interface View : BaseContract.BaseView {
         fun setTextRedeemAmount(redeem: String)
         fun setTextDiscount(discount: String)
-        fun sendDiscountToCheckoutPage(discount: Int)
+        fun sendDiscountToCheckoutPage(discount: BigDecimal)
     }
 
     interface Presenter : BaseContract.BasePresenter<View, BaseContract.BaseCaller> {
-        fun redeemChanged(value: Int, symbol: String)
+        fun redeemChanged(value: BigDecimal, symbol: String)
         fun handleClickRedeem()
     }
 }
